set flag to signal DictionaryCoding is safe for app extensions
Set `APPLICATION_EXTENSION_API_ONLY = YES` so that DictionaryCoding can
be linked from app extensions without seeing the following warning:

```
ld: warning: linking against a dylib which is not safe for use in
  application extensions: ...
```

Under the covers, this toggles some flag on the dynamic library and the
linker checks this flag whenever linking an app extension.

Since DictionaryCoding is not iOS specific, it obviously doesn't use any
iOS specific APIs that are unavailable in app extensions.

Tested that the framework still builds for all platforms and passes
tests.